+++ test/shuriken/namespace_test.clj
@@ -1,5 +1,4 @@
 (ns shuriken.namespace-test
-  (:use clojure.pprint)
   (:require [clojure.test :refer :all]
             [shuriken.core :refer :all
              :reload true]
@@ -49,3 +48,25 @@
     (is (= 'Object
            (-> 'java.lang.Object unqualify fully-qualify unqualify)))))
 
+(deftest test-with-ns
+  (let [initial-ns *ns*]
+    (is (with-ns 'shuriken.virtual-test-namespace
+          (= *ns* (the-ns 'shuriken.virtual-test-namespace))))
+    (is (= *ns* initial-ns))))
+
+(deftest test-once-ns
+  (when (find-ns 'shuriken.once-test-namespace)
+      (clojure.lang.Namespace/remove 'shuriken.once-test-namespace)
+      (dosync (commute @#'clojure.core/*loaded-libs*
+                       disj 'shuriken.once-test-namespace)))
+  (testing "namespace should not be loaded yet"
+    (is (nil? (find-ns 'shuriken.once-test-namespace))))
+  (testing "requiring for the first time"
+    (is (= (with-out-str (require 'shuriken.once-test-namespace))
+           "loaded\n")))
+  (testing "requiring again with :reload"
+    (is (= (with-out-str (require 'shuriken.once-test-namespace :reload))
+           "")))
+  (testing "requiring again with :reload-all"
+    (is (= (with-out-str (require 'shuriken.once-test-namespace :reload-all))
+           ""))))