+++ test/shuriken/once_test_namespace.clj
@@ -1,0 +1,6 @@
+(require 'shuriken.core)
+
+(shuriken.core/once-ns
+  (ns shuriken.once-test-namespace)
+
+  (println "loaded"))